refactor: combine reference resolvers into refs package
This makes it a bit easier to consume, and can enable our Ref objects
to implement an interface for consistent resolution.
