+++ apis/refs/v1beta1/projectref.go
@@ -14,6 +14,19 @@
 
 package v1beta1
 
+import (
+	"context"
+	"fmt"
+	"strings"
+
+	"github.com/GoogleCloudPlatform/k8s-config-connector/pkg/clients/generated/apis/k8s/v1alpha1"
+	apierrors "k8s.io/apimachinery/pkg/api/errors"
+	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
+	"k8s.io/apimachinery/pkg/runtime/schema"
+	"k8s.io/apimachinery/pkg/types"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+)
+
 // The Project that this resource belongs to.
 type ProjectRef struct {
 	/* The `projectID` field of a project, when not managed by KCC. */
@@ -25,4 +38,83 @@ type ProjectRef struct {
 	// The kind of the Project resource; optional but must be `Project` if provided.
 	// +optional
 	Kind string `json:"kind,omitempty"`
+}
+
+// AsProjectRef converts a generic ResourceRef into a ProjectRef
+func AsProjectRef(in *v1alpha1.ResourceRef) *ProjectRef {
+	if in == nil {
+		return nil
+	}
+	return &ProjectRef{
+		Namespace: in.Namespace,
+		Name:      in.Name,
+		External:  in.External,
+		Kind:      in.Kind,
+	}
+}
+
+type Project struct {
+	ProjectID string
+}
+
+// ResolveProject will resolve a ProjectRef to a Project, with the ProjectID.
+func ResolveProject(ctx context.Context, reader client.Reader, src client.Object, ref *ProjectRef) (*Project, error) {
+	if ref == nil {
+		return nil, nil
+	}
+
+	if ref.Kind != "" {
+		if ref.Kind != "Project" {
+			return nil, fmt.Errorf("kind is optional on project reference, but must be \"Project\" if provided")
+		}
+	}
+
+	if ref.External != "" {
+		if ref.Name != "" {
+			return nil, fmt.Errorf("cannot specify both name and external on project reference")
+		}
+
+		tokens := strings.Split(ref.External, "/")
+		if len(tokens) == 1 {
+			return &Project{ProjectID: tokens[0]}, nil
+		}
+		if len(tokens) == 2 && tokens[0] == "projects" {
+			return &Project{ProjectID: tokens[1]}, nil
+		}
+		return nil, fmt.Errorf("format of project external=%q was not known (use projects/<projectId> or <projectId>)", ref.External)
+	}
+
+	if ref.Name == "" {
+		return nil, fmt.Errorf("must specify either name or external on project reference")
+	}
+
+	key := types.NamespacedName{
+		Namespace: ref.Namespace,
+		Name:      ref.Name,
+	}
+	if key.Namespace == "" {
+		key.Namespace = src.GetNamespace()
+	}
+
+	project := &unstructured.Unstructured{}
+	project.SetGroupVersionKind(schema.GroupVersionKind{
+		Group:   "resourcemanager.cnrm.cloud.google.com",
+		Version: "v1beta1",
+		Kind:    "Project",
+	})
+	if err := reader.Get(ctx, key, project); err != nil {
+		if apierrors.IsNotFound(err) {
+			return nil, fmt.Errorf("referenced Project %v not found", key)
+		}
+		return nil, fmt.Errorf("error reading referenced Project %v: %w", key, err)
+	}
+
+	projectID, err := GetResourceID(project)
+	if err != nil {
+		return nil, err
+	}
+
+	return &Project{
+		ProjectID: projectID,
+	}, nil
 }
