+++ apis/refs/v1beta1/resourceid.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package references
+package v1beta1
 
 import (
 	"fmt"
