+++ pkg/controller/direct/privateca/privatecapool_controller.go
@@ -26,10 +26,10 @@ import (
 	"k8s.io/apimachinery/pkg/runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
+	refs "github.com/GoogleCloudPlatform/k8s-config-connector/apis/refs/v1beta1"
 	krm "github.com/GoogleCloudPlatform/k8s-config-connector/pkg/clients/generated/apis/privateca/v1beta1"
 	"github.com/GoogleCloudPlatform/k8s-config-connector/pkg/config"
 	"github.com/GoogleCloudPlatform/k8s-config-connector/pkg/controller/direct/directbase"
-	"github.com/GoogleCloudPlatform/k8s-config-connector/pkg/controller/direct/references"
 	"github.com/GoogleCloudPlatform/k8s-config-connector/pkg/controller/direct/registry"
 )
 
@@ -89,7 +89,7 @@ func (m *caPoolModel) AdapterForObject(ctx context.Context, reader client.Reader
 		return nil, fmt.Errorf("cannot resolve location")
 	}
 
-	projectRef, err := references.ResolveProject(ctx, reader, obj, references.AsProjectRef(&obj.Spec.ProjectRef))
+	projectRef, err := refs.ResolveProject(ctx, reader, obj, refs.AsProjectRef(&obj.Spec.ProjectRef))
 	if err != nil {
 		return nil, err
 	}
